Support for local binapi
VPPTop uses VPP-Agent binary API handlers to communicate with the VPP.
As such, it is limited for VPP versions provided by the Agent and adding
new compatibility is difficult.
This change adds local binary API with simplified API calls
for telemetry, interfaces and VPE, which can be easily updated for any
VPP version, while the original way remained available as well.

* added 'local' with binapi and vppcalls for VPP 20.09
* updated GoVPP to v0.3.5
* Minor fixes and cleanup

Signed-off-by: Vladimir Lavor <vlavor@cisco.com>