Local-only VPPTop
The goal of this change is to allow running VPPTop using only
local VPP implementation as an addition to the full VPP support,
and enable unbind it from the VPPAgent.
Repository now contains two mains - the one called vpptop works
as before, the second one called vpptop_local uses only the local
implementation and is not dependent on the VPPAgent.

* /cmd folder now contains two binaries for vanilla vpptop and for
  the local one. Command-related files previously situated there
  were moved to new /command directory.
* Every VPP handler (Ligato agent-based and local-based) has its own
  definiton struct which enables the handler in the VPPTop and
  validates compatibility with the connected VPP version.
* Added HandlerAPI as a common handler interface designed for
  VPP provider (which is responsible for collecting stats).
* Added ProviderAPI designed for VPPTop app for easier and better
  readable access.

Other improvements:
* Added simple Makefile to easier build and binary API generation
* GoVPP updated to v0.3.5
* Fixed node (runtime) readout for local impl.
* Some table dimensions were adjusted to better show displayed data
* Various minor code improvements

TODO: docs

Signed-off-by: Vladimir Lavor <vlavor@cisco.com>